pyproject.toml
[project]
name = "migration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio-throttle>=1.0.2",
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "google-cloud-storage>=3.3.0",
    "pydantic>=2.11.7",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.1.1",
    "supabase>=2.18.1",
    "tqdm>=4.67.1",
    "turbopuffer>=1.0.0",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "ruff>=0.12.9",
]
